// angular-frontend/src/app/models/interfaces.ts
export interface Capteur {
    id: number;
    type: 'HUMIDITE' | 'TEMPERATURE' | 'LUMINOSITE';
    parcelleId: number;
    emplacement: string;
}

export interface Observation {
    id: number;
    capteurId: number;
    valeur: number;
    unite: string;
    date: string;
}

export interface Modele {
    id: number;
    type: string;
    version: string;
    dateApprentissage: string;
    precision: number;
}

export interface Recommandation {
    id: number;
    parcelleId: number;
    quantiteEau: number;
    justification: string;
    dateCalcul?: string; // Optional as it might be generated by backend
}